+++ src/main/java/com/project/kisan_setu/dto/DashboardDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,5 +13,7 @@ public class DashboardDto {
     private Long activeListings;
     private Long totalBidsReceived;
     private Long pendingApprovals;
-    private Long totalRevenue;
+    private BigDecimal totalRevenue;
+
+
 }
