+++ src/main/java/com/project/kisan_setu/entity/Listing.java
@@ -6,12 +6,14 @@ import com.project.kisan_setu.enums.AuctionStatus;
 import com.project.kisan_setu.enums.PurchaseType;
 import com.project.kisan_setu.enums.SaleType;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -44,10 +46,10 @@ public class Listing {
 
     // Pricing & Purchase Type
     private Double pricePerKg;
-    private Double totalBasePrice;
+    private BigDecimal totalBasePrice;
     @Enumerated(EnumType.STRING)
     private SaleType saleType;
-    private Double minimumBidIncrement;
+    private BigDecimal minimumBidIncrement;
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt;
 
@@ -99,6 +101,7 @@ public class Listing {
     // Relationship with buying requirements
     @OneToMany(mappedBy = "listing", cascade = CascadeType.ALL)
     private List<BuyingRequirement> buyingRequirements = new ArrayList<>();
+
